# Reduce excessive logging in debug.log and perf.log