--- a/OpenRA.Game/Game.cs
+++ b/OpenRA.Game/Game.cs
@@ -606,8 +606,6 @@ namespace OpenRA
 					{
 						++orderManager.LocalFrameNumber;
 
-						Log.Write("debug", "--Tick: {0} ({1})", LocalTick, isNetTick ? "net" : "local");
-
 						if (isNetTick)
 							orderManager.Tick();