--- a/mods/common/chrome/settings-advanced.yaml
+++ b/mods/common/chrome/settings-advanced.yaml
@@ -115,3 +115,10 @@ Container@ADVANCED_PANEL:
 					Height: 20
 					Font: Regular
 					Text: Enable Debug Commands in Replays
+				Checkbox@PERFLOGGING_CHECKBOX:
+					X: 310
+					Y: 273
+					Width: 300
+					Height: 20
+					Font: Regular
+					Text: Enable Tick Performance Logging