Following https://github.com/OpenRA/OpenRA/issues/19029#issuecomment-803683420, this PR is meant to complement #19264 and/or other solutions to hopefully solve #19260 for good.

While having as much information as possible available to debug any crashes is generally valuable, the amount of of logging we do seems excessive for the purpose it's mean to serve.

Logging the `--Tick` in `debug.log` seems mostly pointless, as the few kinds of potential crash sources getting logged in debug.log either happen at/before game start, or can include the world tick in their log entry.

Meanwhile, the perf.log of players rarely shows us devs anything we don't already know (except infinite loops when their cost exceeds the threshold, but that's a poor excuse to drag down performance for players).

Additionally, in both cases developers can still set their preferred value in settings.yaml and run a provided replay to get the needed information, if necessary.