@dnqbob 2021-03-22
I think CastleJing will like this idea and he planned to do similar change to TA-engine