@dnqbob 2021-03-22
Maybe use a loop queue to record last few ticks?