@reaperrr 2021-03-22
> Making this opt-in is a good idea, but changing the default to 10ms makes the feature essentially useless. The idea of this logging was to highlight the intermediate cases that aren't bad enough to cause an obvious lag spike on their own, but when combined with other intermediate cases in the same tick still completely blow the frame budget. We already know and mostly can't fix the 10ms+ cases, so don't care about those!

Point taken. But making the perf-logging opt-in as well would be fine? Or at least making it opt-out?