@pchote 2021-03-22
IMO we should treat it the same way we do `SyncCheckUnsyncedCode`: disabled by default, with a checkbox hidden in the Developer Settings menu.