@reaperrr 2021-03-22
OK, though that would make us circle back to the lack of space in the Advanced settings, especially once #19242 is merged as well.

Would squeezing the debug checkboxes closer together (vertically) be acceptable until the menu is restructured?