@pchote 2021-03-22
#19242 won't work as a debug option (see my comment there), which solves the issue here 🙂