@reaperrr 2021-03-26
Before I update this, some actual performance numbers.
I measured by logging the average of the last 500 ticks (via PerfHistory) at ticks 501 and 1001.
Also did 5 runs each with and without logging enabled, to reduce impact of fluctuations from CPU clockspeeds or background tasks.

RA shellmap with perf logging enabled, like on bleed:
```
Run 1:
--Tick: 501, 2,649339393939394 ms
--Tick: 1001, 2,6562929292929276 ms

Run 2:
--Tick: 501, 2,6692050505050493 ms
--Tick: 1001, 2,664829292929293 ms

Rund 3:
--Tick: 501, 2,6916747474747464 ms
--Tick: 1001, 2,684061616161616 ms

Run 4:
--Tick: 501, 2,670688888888888 ms
--Tick: 1001, 2,673647474747474 ms

Run 5:
--Tick: 501, 2,651198989898989 ms
--Tick: 1001, 2,673625252525253 ms
```

With perf logging (including the GetTimestamp calls) disabled:
```
Run 1:
--Tick: 501, 2,4912707070707074 ms
--Tick: 1001, 2,470488888888889 ms

Run 2:
--Tick: 501, 2,424936363636364 ms
--Tick: 1001, 2,4316414141414135 ms

Run 3:
--Tick: 501, 2,584616161616162 ms
--Tick: 1001, 2,498614141414141 ms

Run 4:
--Tick: 501, 2,42079191919192 ms
--Tick: 1001, 2,4808181818181816 ms

Run 5:
--Tick: 501, 2,5107303030303028 ms
--Tick: 1001, 2,544620202020202 ms
```

Looks like about 5-7% less tick time on the RA shellmap just from disabling perf logging, and we're talking about a Ryzen 3600 here, on slower systems + larger games with significantly more perf logging we'd likely see a 2-digit performance impact.

Of course, #19264 will likely considerably reduce (or even eradicate?) the impact of this PR.