@reaperrr 2021-03-26
> which solves the issue here 🙂

Technically, I'd need 2 checkboxes, unless we want to merge the options into a generic "Additional debug/perf logging" option.
For now I only exposed the perf logging option, since I consider that one more important.