@dnqbob 2021-03-26
I will suggest some mods that already slowed by trait, map and AI use log record fix PR