@reaperrr 2021-03-28
Updated.

Also pulled out `var`s for the debug option to make things a bit more readable (maybe saving a microsecond per lookup as well).