@teinarss 2021-04-03
We have the same logic in 3 places and it also makes the code a bit harder to read, one idea is to move the logic to a dedicated class see https://github.com/teinarss/OpenRA/commit/cd363e35a08befb4c184898303f026fb1d2f9a12 for a example how we could do it. We could also let the ShouldLog method be private and let Log call it.