@reaperrr 2021-04-04
Updated.

Compared to @teinarss code, I merged the ShouldLog() method into Log(), since the actual logging must happen before `start` is reset so we can't do this in separate methods (unless we cache it, but I felt just merging the methods is a simpler solution).