@reaperrr 2021-04-10
Updated. Used the opportunity to rename the logger to better match the other Perf* measure helpers.